# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Backend Pleno - Soluct API",
        "description": "Documentação da API RESTful para o desafio técnico da Soluct.\n * * **Observações Importantes:**\n * - Todas as requisições que enviam dados no corpo (POST, PUT, PATCH) devem usar o header `Content-Type: application/json`.\n * - Todas as requisições devem incluir o header `Accept: application/json` para garantir que a resposta seja recebida no formato JSON correto.\n * - Os endpoints que requerem autenticação estão marcados com um ícone de cadeado.",
        "contact": {
            "email": "elizaelnycolas@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Listar as tarefas do usuário",
                "description": "Retorna uma lista paginada de tarefas pertencentes ao usuário autenticado. Este endpoint suporta filtros complexos, ordenação e paginação.\n *\n * **Exemplo de Uso Completo:**\n * Para buscar tarefas com status 'pending' ou 'in_progress', com data de vencimento entre 2025-01-01 e 2025-01-31, ordenadas pela data de vencimento (da mais antiga para a mais nova), a URL seria:\n * `/api/v1/tasks?filters[status][operator]=in&filters[status][value]=pending,in_progress&filters[due_date][operator]=between&filters[due_date][value]=2025-01-01,2025-01-31&sort_by=due_date&sort_order=asc`",
                "operationId": "860ebffacd8213547f9aeb5927f133dd",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de itens por página.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 500
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Campo para ordenação. Colunas permitidas: id, title, status, due_date, created_at.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Ordem da ordenação.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "filters[status][value]",
                        "in": "query",
                        "description": "Filtra tarefas por status. Para múltiplos status, separe por vírgula (ex: pendente,em_andamento).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters[status][operator]",
                        "in": "query",
                        "description": "Operador para o filtro de status.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "=",
                            "enum": [
                                "=",
                                "in"
                            ]
                        }
                    },
                    {
                        "name": "filters[title][value]",
                        "in": "query",
                        "description": "Filtra tarefas por título (busca parcial).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters[title][operator]",
                        "in": "query",
                        "description": "Operador para o filtro de título.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "like",
                            "enum": [
                                "like",
                                "="
                            ]
                        }
                    },
                    {
                        "name": "filters[due_date][value]",
                        "in": "query",
                        "description": "Filtra por data de vencimento. Para o operador 'between', use duas datas separadas por vírgula (YYYY-MM-DD,YYYY-MM-DD).",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters[due_date][operator]",
                        "in": "query",
                        "description": "Operador para o filtro de data de vencimento.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "=",
                            "enum": [
                                "=",
                                ">",
                                "<",
                                ">=",
                                "<=",
                                "between"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida. Retorna a lista paginada de tarefas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Título da Tarefa"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Descrição da Tarefa"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "due_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-12-31"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/v1/tasks?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/v1/tasks?page=5"
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://localhost/api/v1/tasks?page=2",
                                            "nullable": true
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost/api/v1/tasks"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 75
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Cria uma nova tarefa",
                "operationId": "97c5a57392007d72f65fe0d21c5e1a45",
                "requestBody": {
                    "description": "Dados da tarefa",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "completed",
                                            "canceled"
                                        ]
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarefa criada com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Exibe uma tarefa",
                "operationId": "1283a9795ea696666cf34f35300fa531",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da tarefa"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Atualiza uma tarefa",
                "operationId": "064ac3d4f1ad4022c01b9ef7d9fd84c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados para atualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "enumeration",
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "completed"
                                        ]
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarefa atualizada com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Deleta uma tarefa",
                "operationId": "56771919c3b08ef9c4128f60f35837d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tarefa deletada com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}